app/requirements.txt
langchain
langchain-openai
langchain-community
langchain-chroma
openai
python-dotenv
PyMuPDF
chromadb
tiktoken
streamlit
langfuse>=2.0.0
pandas
matplotlib
tavily-python